requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
transformers>=4.38.0
open-clip-torch>=2.24.0
albumentations>=1.4.0
opencv-python>=4.9.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
Pillow>=10.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
pyyaml>=6.0
pandas>=2.0.0
seaborn>=0.12.0
